+++ src/Driver/ServerGoneAwayExceptionsAwareTrait.php
@@ -2,6 +2,8 @@
 
 namespace Facile\DoctrineMySQLComeBack\Doctrine\DBAL\Driver;
 
+use Exception;
+
 /**
  * Trait ServerGoneAwayExceptionsAwareTrait.
  */
@@ -19,11 +21,11 @@ trait ServerGoneAwayExceptionsAwareTrait
     ];
 
     /**
-     * @param \Exception $exception
+     * @param Exception $exception
      *
      * @return bool
      */
-    public function isGoneAwayException(\Exception $exception)
+    public function isGoneAwayException(Exception $exception): bool
     {
         $message = $exception->getMessage();
 
@@ -37,11 +39,11 @@ trait ServerGoneAwayExceptionsAwareTrait
     }
 
     /**
-     * @param \Exception $exception
+     * @param Exception $exception
      *
      * @return bool
      */
-    public function isGoneAwayInUpdateException(\Exception $exception)
+    public function isGoneAwayInUpdateException(Exception $exception): bool
     {
         $message = $exception->getMessage();