+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.9.0] - TBD
+### Added
+ * Added compatibility with doctrine/dbal 2.11
+ * Added Github Actions for CI
+### Changed
+ * Bumped minimum PHP version to 7.3
+ * Updated dependencies
+ * Added functional tests
+### Fixed
+ * Fixed compatiblity with doctrine/dbal 2.11
+
 ## [1.8] - 2019-09-05
 ### Fixed
  * Fixed issue about loss of state for Statement class on retry (#34)